+++ Modules/Ruleset.ProgramInfo/Public/Test-FileSystemPath.ps1
@@ -173,12 +173,12 @@ function Test-FileSystemPath
 		if ($Quiet)
 		{
 			# Make sure -Quiet switch does not make troubleshooting hard
-			Write-Debug -Message "[$InvocationName & WriteConditional] $Message"
+			Write-Debug -Message "[$InvocationName - WriteConditional] $Message"
 		}
 		else
 		{
-			Write-Warning -Message "[$InvocationName & WriteConditional] $Message"
-			Write-Information -Tags "$InvocationName & WriteConditional" -MessageData "INFO: Path '$LiteralPath'"
+			Write-Warning -Message "[$InvocationName - WriteConditional] $Message"
+			Write-Information -Tags "$InvocationName - WriteConditional" -MessageData "INFO: Path '$LiteralPath'"
 		}
 	}