+++ Scripts/Deploy-Firewall.ps1
@@ -28,7 +28,7 @@ SOFTWARE.
 
 <#PSScriptInfo
 
-.VERSION 0.13.0
+.VERSION 0.13.1
 
 .GUID d7809432-c822-4699-9244-97c8da8d64bf
 
@@ -84,7 +84,6 @@ None. Deploy-Firewall.ps1 does not generate any output
 
 .NOTES
 TODO: Rule deployment should probably be separated into new script
-TODO: Quiet switch should not suppress errors ex. about failed digital signature
 
 .LINK
 https://github.com/metablaster/WindowsFirewallRuleset/tree/master/Scripts