+++ Scripts/Utility/Restart-Network.ps1
@@ -28,7 +28,7 @@ SOFTWARE.
 
 <#PSScriptInfo
 
-.VERSION 0.13.0
+.VERSION 0.13.1
 
 .GUID 51f98c20-9f7b-4be2-bb16-86c76c883dda